+++ src/Order/CheckoutForm.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace SwipeStripe\Forms;
+namespace SwipeStripe\Order;
 
 use SilverStripe\Control\HTTPResponse;
 use SilverStripe\Control\RequestHandler;
@@ -22,14 +22,12 @@ use SilverStripe\Omnipay\Model\Message\PurchaseError;
 use SilverStripe\Omnipay\Model\Payment;
 use SilverStripe\Omnipay\Service\ServiceFactory;
 use SilverStripe\Security\Security;
-use SwipeStripe\Order\Order;
-use SwipeStripe\Order\PaymentStatus;
 use SwipeStripe\Pages\OrderConfirmationPage;
 use SwipeStripe\Price\SupportedCurrencies\SupportedCurrenciesInterface;
 
 /**
  * Class CheckoutForm
- * @package SwipeStripe\Forms
+ * @package SwipeStripe\Order
  * @property Payment|null $PaymentError
  */
 class CheckoutForm extends Form
